src/pyproject.toml
[project]
name = "agent-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncio-mqtt>=0.11.0",
    "dataclasses-json>=0.6.0",
    "httpx>=0.25.0",
    "langchain>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-core>=0.2.0",
    "langchain-deepseek>=0.1.4",
    "langchain-google-genai>=1.0.0",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-mistralai>=0.2.11",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.1.0",
    "mcp>=1.0.0",
    "nest-asyncio>=1.6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
    "toml>=0.10.0",
    "typing-extensions>=4.8.0",
]
